requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
openai==1.47.0
tiktoken==0.7.0
chromadb==0.5.5
typer==0.12.5
rich==13.8.1
streamlit==1.32.0
plotly==5.18.0
supabase==2.3.4
pgvector==0.2.4
